+++ test/src/lib.rs
@@ -1,8 +1,3 @@
-#![cfg_attr(test, feature(test))]
-
-#[cfg(test)]
-extern crate test;
-
 #[macro_use]
 extern crate serde_derive;